refactor: cleanup file locations, add a lot of documentation
