+++ src/lib/mappers/abilityMapper.ts
@@ -1,7 +1,7 @@
 import { pokedex } from '#assets/pokedex';
 import type { PokemonTypes } from '#assets/pokemon-source';
 import { mapPokemonDataToPokemonGraphQL, PokemonReferencedCallIdentifier } from '#mappers/pokemonMapper';
-import type { Ability, Pokemon } from '#types';
+import type { Ability, Pokemon } from '#types/graphql-mapped-types';
 import { addPropertyToObjectFieldBased } from '#utils/addPropertyToObject';
 import type { GraphQLSet } from '#utils/GraphQLSet';
 import { isNullishOrEmpty, toTitleCase } from '@sapphire/utilities';
