+++ src/lib/mappers/moveMapper.ts
@@ -1,8 +1,8 @@
 import type { PokemonTypes } from '#assets/pokemon-source';
-import type { Move } from '#types';
-import { addPropertyToObjectFieldBased } from '#utils/addPropertyToObject';
+import type { Move } from '#types/graphql-mapped-types';
 import type { GraphQLSet } from '#utils/GraphQLSet';
-import { parseZCrystal, toLowerHyphenCase, toLowerSingleWordCase, toTitleSnakeCaseWithoutHyphenConvert } from '#utils/util';
+import { addPropertyToObjectFieldBased } from '#utils/addPropertyToObject';
+import { parseZCrystal, toLowerHyphenCase, toLowerSingleWordCase, toTitleSnakeCaseWithoutHyphenConvert } from '#utils/utils';
 
 export function mapMoveDataToMoveGraphQL({ data, requestedFields }: MapMoveDataToMoveGraphQLParameters): Move {
   const moveObject: Move = {} as Move;
