+++ src/lib/utils/addPropertyToObject.ts
@@ -1,6 +1,14 @@
 import type { GraphQLSet } from '#utils/GraphQLSet';
 import { isFunction, isNullish } from '@sapphire/utilities';
 
+/**
+ * Adds a property to an object conditionally based on a given condition.
+ *
+ * @template T - The type of the object.
+ * @template K - The type of the property key.
+ * @param {AddPropertyToObjectConditionalParameters<T, K>} params - The parameters for adding the property.
+ * @returns The object with the property added conditionally.
+ */
 export function addPropertyToObjectConditional<T extends object, K extends keyof T>({
   objectTarget,
   propertyKey,
@@ -18,6 +26,15 @@ export function addPropertyToObjectConditional<T extends object, K extends keyof
   return objectTarget;
 }
 
+/**
+ * Adds a property to an object based on the requested fields.
+ * If the requested fields include the field accessor, the property is added to the object.
+ *
+ * @template T - The type of the object.
+ * @template K - The type of the property key.
+ * @param {AddPropertyToObjectParameters<T, K>} params - The parameters for adding the property.
+ * @returns The object with the property added.
+ */
 export function addPropertyToObjectFieldBased<T extends object, K extends keyof T>({
   objectTarget,
   propertyKey,
@@ -36,6 +53,13 @@ export function addPropertyToObjectFieldBased<T extends object, K extends keyof 
   return objectTarget;
 }
 
+/**
+ * Adds a property to an object.
+ *
+ * @template T - The type of the object.
+ * @template K - The key of the property to be added.
+ * @param params - The `objectTarget`, `propertyKey`, and `propertyValue` parameters.
+ */
 function addPropertyToObject<T extends object, K extends keyof T>({
   objectTarget,
   propertyKey,
@@ -48,6 +72,11 @@ function addPropertyToObject<T extends object, K extends keyof T>({
   }
 }
 
+/**
+ * Represents the conditional parameters for adding a property to an object.
+ * @template T - The type of the object.
+ * @template K - The key of the property to be added.
+ */
 type AddPropertyToObjectConditionalParameters<T extends object, K extends keyof T> = Omit<
   AddPropertyToObjectParameters<T, K>,
   'requestedFields' | 'fieldAccessor'
@@ -55,6 +84,11 @@ type AddPropertyToObjectConditionalParameters<T extends object, K extends keyof 
   condition: boolean;
 };
 
+/**
+ * Represents the parameters for adding a property to an object.
+ * @template T - The type of the object.
+ * @template K - The type of the property key.
+ */
 interface AddPropertyToObjectParameters<T extends object, K extends keyof T> {
   objectTarget: T;
   propertyKey: K;
