+++ src/lib/utils/pokemonTypes.ts
@@ -1,20 +1,39 @@
+/** Enum representing the different types of Pokémon. */
 export enum TypesEnum {
+  /** Represents the Bug type. */
   Bug = 'bug',
+  /** Represents the Dark type. */
   Dark = 'dark',
+  /** Represents the Dragon type. */
   Dragon = 'dragon',
+  /** Represents the Electric type. */
   Electric = 'electric',
+  /** Represents the Fairy type. */
   Fairy = 'fairy',
+  /** Represents the Fighting type. */
   Fighting = 'fighting',
+  /** Represents the Fire type. */
   Fire = 'fire',
+  /** Represents the Flying type. */
   Flying = 'flying',
+  /** Represents the Ghost type. */
   Ghost = 'ghost',
+  /** Represents the Grass type. */
   Grass = 'grass',
+  /** Represents the Ground type. */
   Ground = 'ground',
+  /** Represents the Ice type. */
   Ice = 'ice',
+  /** Represents the Normal type. */
   Normal = 'normal',
+  /** Represents the Poison type. */
   Poison = 'poison',
+  /** Represents the Psychic type. */
   Psychic = 'psychic',
+  /** Represents the Rock type. */
   Rock = 'rock',
+  /** Represents the Steel type. */
   Steel = 'steel',
+  /** Represents the Water type. */
   Water = 'water'
 }
