+++ src/tsconfig.json
@@ -2,19 +2,7 @@
   "extends": "../tsconfig.base.json",
   "compilerOptions": {
     "rootDir": "./",
-    "outDir": "../api",
-    "baseUrl": ".",
-    "paths": {
-      "#assets/*": ["./lib/assets/*"],
-      "#jsonAssets/*.json": ["./lib/assets/*.json"],
-      "#dexdata/*": ["./lib/assets/pokedex-data/*"],
-      "#resolvers/*": ["./lib/resolvers/*"],
-      "#types": ["./lib/mapped-types/graphql-mapped-types.d.ts"],
-      "#utils/*": ["./lib/utils/*"],
-      "#validations/*": ["./lib/validations/*"],
-      "#mappers/*": ["./lib/mappers/*"],
-      "#root/*": ["*"]
-    }
+    "outDir": "../api"
   },
   "include": [".", "./lib/assets/**/*.json"]
 }
