+++ tests/tsconfig.json
@@ -7,7 +7,7 @@
       "#jsonAssets/*.json": ["../src/lib/assets/*.json"],
       "#dexdata/*": ["../src/lib/assets/pokedex-data/*"],
       "#resolvers/*": ["../src/lib/resolvers/*"],
-      "#types": ["../src/lib/mapped-types/graphql-mapped-types.d.ts"],
+      "#types/*": ["../src/lib/types/*"],
       "#utils/*": ["../src/lib/utils/*"],
       "#validations/*": ["../src/lib/validations/*"],
       "#mappers/*": ["../src/lib/mappers/*"],
@@ -16,5 +16,5 @@
     },
     "types": ["vitest/globals"]
   },
-  "include": [".", "../src/lib/mapped-types"]
+  "include": ["."]
 }
